src/aws-finops-mcp-server/pyproject.toml
[project]
name = "awslabs.aws-finops-mcp-server"

# NOTE: "Patch"=9223372036854775807 bumps next release to zero.
version = "0.0.0"

description = "A Model Context Protocol (MCP) server that provides tools for AWS cost optimization by wrapping boto3 SDK functions."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "loguru>=0.7.0",
    "mcp[cli]>=1.6.0",
    "pydantic>=2.10.6",
    "fastmcp>=2.0.0",
    "boto3>=1.28.0",
    "python-dotenv>=1.0.0",
]
license = {text = "Apache-2.0"}
license-files = ["LICENSE", "NOTICE" ]
authors = [
    {name = "Amazon Web Services"},
    {name = "AWSLabs MCP", email="203918161+awslabs-mcp@users.noreply.github.com"},
    {name = "Rams Sundaram", email="shsrams@amazon.com"},
]
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

[project.urls]
homepage = "https://awslabs.github.io/mcp/"
docs = "https://awslabs.github.io/mcp/servers/aws-finops-mcp-server/"
documentation = "https://awslabs.github.io/mcp/servers/aws-finops-mcp-server/"
repository = "https://github.com/awslabs/mcp.git"
changelog = "https://github.com/awslabs/mcp/blob/main/src/aws-finops-mcp-server/CHANGELOG.md"

[project.scripts]
"awslabs.aws-finops-mcp-server" = "awslabs.aws_finops_mcp_server.server:main"

[dependency-groups]
dev = [
    "commitizen>=4.2.2",
    "pre-commit>=4.1.0",
    "ruff>=0.9.7",
    "pyright>=1.1.398",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff]
line-length = 99
extend-include = ["*.ipynb"]
exclude = [
    ".venv",
    "**/__pycache__",
    "**/node_modules",
    "**/dist",
    "**/build",
    "**/env",
    "**/.ruff_cache",
    "**/.venv",
    "**/.ipynb_checkpoints"
]
force-exclude = true

[tool.ruff.lint]
exclude = ["__init__.py"]
select = ["C", "D", "E", "F", "I", "W"]
ignore = ["C901", "E501", "E741", "F402", "F823", "D100", "D106"]

[tool.ruff.lint.isort]
lines-after-imports = 2
no-sections = true

[tool.ruff.lint.per-file-ignores]
"**/*.ipynb" = ["F704"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true

[tool.pyright]
include = ["awslabs", "tests"]
exclude = ["**/__pycache__", "**/.venv", "**/node_modules", "**/dist", "**/build"]

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.0.0"
tag_format = "v$version"
version_files = [
    "pyproject.toml:version",
    "awslabs/aws_finops_mcp_server/__init__.py:__version__"
]
update_changelog_on_bump = true

[tool.hatch.build.targets.wheel]
packages = ["awslabs"]

[tool.bandit]
exclude_dirs = ["venv", ".venv", "tests"]

[tool.pytest.ini_options]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
testpaths = ["tests"]
asyncio_mode = "auto"
log_cli = true
log_cli_level = "INFO"
markers = [
    "integration: marks tests as integration tests that make actual AWS API calls",
    "unit: marks tests as unit tests that don't make external calls",
    "live: marks tests that make live API calls (deselect with '-m \"not live\"')",
    "asyncio: marks tests that use asyncio"
]

[tool.coverage.report]
exclude_also = [
    'pragma: no cover',
    'if __name__ == .__main__.:\n    main()',
]

[tool.coverage.run]
source = ["awslabs"]
